Fix compiler warnings for cmake build
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@166707 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a52f93ddd0d26c534c430eaf0b36586b513c12c5

+++ Online/Online/Gaucho/src/MonHist.cpp
@@ -650,7 +650,7 @@ int MonHist::serialize(void* &ptr)
   DimBuffBase *pill = (DimBuffBase*)ptr;
   if(m_type == H_ILLEGAL)
   {
-    pill->type = H_ILLEGAL;
+    pill->type = (int)H_ILLEGAL;
     pill->reclen = sizeof(*pill);
     pill->nameoff=sizeof(*pill);
     pill->namelen = 0;
